refactor: Swagger 어노테이션 누락 정리 및 adapter/in/web/ 패키지 구조 정리
