+++ src/main/java/com/roovies/concertreservation/balance/infra/adapter/in/web/dto/response/GetBalanceResponse.java
@@ -1,0 +1,17 @@
+package com.roovies.concertreservation.balance.infra.adapter.in.web.dto.response;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.Builder;
+
+import java.time.LocalDateTime;
+
+@Builder
+@Schema(description = "잔액 조회 응답 DTO")
+public record GetBalanceResponse(
+        @Schema(description = "보유 잔액", example = "20000")
+        Long balance,
+
+        @Schema(description = "응답 시간", example = "2025-08-26T21:00:00")
+        LocalDateTime responseTime
+) {
+}
