--- src/main/java/com/roovies/concertreservation/balance/infra/adapter/in/web/request/UpdateBalanceRequest.java
@@ -1,6 +1,0 @@
-package com.roovies.concertreservation.balance.infra.adapter.in.web.request;
-
-public record UpdateBalanceRequest(
-        Long balance
-) {
-}
