--- src/main/java/com/roovies/concertreservation/balance/infra/adapter/in/web/response/GetBalanceResponse.java
@@ -1,12 +1,0 @@
-package com.roovies.concertreservation.balance.infra.adapter.in.web.response;
-
-import lombok.Builder;
-
-import java.time.LocalDateTime;
-
-@Builder
-public record GetBalanceResponse(
-        Long balance,
-        LocalDateTime responseTime
-) {
-}
