+++ src/main/java/com/roovies/concertreservation/reservations/infra/adapter/in/web/dto/response/GetAvailableSchedulesResponse.java
@@ -7,6 +7,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Builder
+@Schema(description = "콘서트 예약 가능 일정 목록 조회 응답 DTO")
 public record GetAvailableSchedulesResponse(
         @Schema(description = "콘서트 ID", example = "101")
         Long concertId,
